NeighborlyAPI/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.19.0
pymongo==4.6.3
flask==1.1.4
requests==2.31.0
flask-bootstrap==3.3.7.1
flask-restplus==0.13.0
flask-swagger-ui==4.11.1
werkzeug==0.16.1
azure-cosmos==4.6.0
azure-cli-core==2.59.0
feedgen==1.0.0
certifi==2024.2.2
